Make the base64 encoding / decoding functions public
It would be preferable to keep them private, but right now we have to keep duplicates in other packages, so making them public will reduce overall code size.
