+++ packages/bench-codesize/README.md
@@ -9,5 +9,5 @@ minify the bundle, and compress it like a web server would usually do.
 
 | code generator    | bundle size             | minified               | compressed         |
 |-------------------|------------------------:|-----------------------:|-------------------:|
-| protobuf-es       | 160,471 b      | 86,418 b | 15,128 b |
+| protobuf-es       | 160,617 b      | 86,421 b | 15,136 b |
 | google-protobuf   | 368,034 b  | 270,748 b | 43,704 b |
